creating RenderActiveShape.jsx file
